Update to 1.18.2
Removed the tag grouping option. Not sure if this makes sense anymore to keep. with the all the tags that exist now. For example this would consider bamboo and all types of dirt as the same block since they are all in "bamboo_plantable_on"

Custom Grouping now has support for tags. THis should be used to replace the old tag grouping option.